SAM rebalance (circle radius, faster speed)
First missile are too slow sometimes which is weird when they don't
catch up with missile so I increased speed. Made the range a circle. The
guy who made the PR to show the radius of SAM told me radius needs to
decrease compared to "searchRange" as circles is bigger than rectangles
otherwise (on some part of the circle)

Instead of 10% hit made it a 50% hit. The reasoning is as follow: sam
counter is forcing a cooldown (7.5s) by launching a nuke. If player
fails to counter sam with a nuke then it's fair it's H bomb has 50%
chance of dying. If it's only 10% then SAM is totally useless as almost
any H bomb will go through.